:recycle: Rename mod
